refactor(settings): extract save layer into useSettingsSaveHandlers — logic 100%
Move profile/email/password/content-prefs/backup-email saves verbatim
into _hooks/use-settings-save-handlers.ts (owns its own auth + Settings
i18n; forms threaded in since the render binds them), and the 4 zod
schemas into _lib/schemas.ts so page (useForm/zodResolver) and hook
(inferred types) share them without a cycle. Pruned now-unused z/
isValidPassword/supabase imports and the moved useAuth methods.

settings/page is now logic-complete: every save/account/verify/pref/
avatar concern lives in a hook; only render-coupled effects + the
render remain. Behavior-identical; page.tsx 1430->1303 LOC
(1687->1303, -23% across the sweep).
